Rename validator route and add outline helper
